remove unneeded Recoil stuff
